Unified naming convention for PropHunt.Mixed.Utility -> PropHunt.Mixed.Utils
