+++ Assets/Scripts/Mixed/Utils/FilteringClosestHitCollector.cs
@@ -1,12 +1,10 @@
+using System;
+using System.Linq;
 using Unity.Physics;
 
-namespace PropHunt.Mixed.Utilities
+namespace PropHunt.Mixed.Utils
 {
-    /// <summary>
-    /// Collider processor to handle filtering out an object to stop
-    /// hitting itself.
-    /// </summary>
-    public struct SelfFilteringClosestHitCollector<T> : ICollector<T> where T : struct, IQueryResult
+    public struct FilteringClosestHitCollector<T> : ICollector<T> where T : struct, IQueryResult
     {
         /// <summary>
         /// Can this exit early on the current hit object
@@ -33,23 +31,30 @@ namespace PropHunt.Mixed.Utilities
         /// <summary>
         /// Entity index to avoid collisions with.
         /// </summary>
-        private readonly int selfEntityIndex;
+        private readonly int ignore;
+
+        /// <summary>
+        /// Entity index to allow collisions with
+        /// </summary>
+        private readonly int allow;
 
         private CollisionWorld collisionWorld;
 
         /// <summary>
         /// Creates a self filtering object collision detection
         /// </summary>
-        /// <param name="entityId">Index of this entity</param>
+        /// <param name="ignore">index of entity to ignore hitting</param>
+        /// <param name="allow">index of entity to allow to hit, leave empty to allow everything (except the ignroe)</param>
         /// <param name="maxFraction">Maximum fraction that an object can be encountered
         /// as a portion of the current raycast draw</param>
         /// <param name="collisionWorld">World of all colliable objects</param>
-        public SelfFilteringClosestHitCollector(int entityIndex, float maxFraction, CollisionWorld collisionWorld)
+        public FilteringClosestHitCollector(int ignore, int allow, float maxFraction, CollisionWorld collisionWorld)
         {
             this.MaxFraction = maxFraction;
             this.ClosestHit = default;
             this.NumHits = 0;
-            this.selfEntityIndex = entityIndex;
+            this.ignore = ignore;
+            this.allow = allow;
             this.collisionWorld = collisionWorld;
         }
 
@@ -58,7 +63,10 @@ namespace PropHunt.Mixed.Utilities
         /// <inheritdoc/>
         public bool AddHit(T hit)
         {
-            if (collisionWorld.Bodies[hit.RigidBodyIndex].Entity.Index == this.selfEntityIndex)
+            int hitIndex = collisionWorld.Bodies[hit.RigidBodyIndex].Entity.Index;
+            // Ignore if index is part of hit index or if
+            // allowed is defined and does not contain this index
+            if (ignore == hitIndex || (allow != AllowAll && allow != hitIndex))
             {
                 return false;
             }
@@ -69,7 +77,10 @@ namespace PropHunt.Mixed.Utilities
         }
 
         #endregion
+
+        /// <summary>
+        /// Empty set of integers
+        /// </summary>
+        public static readonly int AllowAll = -1;
     }
-
-
 }
